requirements.txt
click==8.1.8
dlib==19.24.8
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.8.0.post1
numpy==1.26.3
opencv-python==4.11.0.86
packaging==25.0
pillow==10.4.0
